Implementazione e testing del metodo sommaMedaglieDiAtletiConSportChiuso

+++ Compito28Aprile/src/it/prova/atletasport/dao/AtletaDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface AtletaDAO extends IBaseDAO<Atleta>{
 
-    public int sumMedaglieBySportChiuso() throws Exception;
+    public Long sumMedaglieBySportChiuso() throws Exception;
 
     public List<Atleta> findAllBySportChiuso() throws Exception;
